+++ R/stat_contour_fill.R
@@ -140,7 +140,10 @@ StatContourFill <- ggplot2::ggproto("StatContourFill", ggplot2::Stat,
 
 
         # Make contours
+        dec <- getOption("OutDec")
+        options(OutDec = ".")
         cont <- data.table::setDT(.contour_bands(data, breaks, complete = complete, clip = clip, proj = proj))
+        options(OutDec = dec)
 
         cont[, int.level := (level_high + level_low)/2]
         cont[, level_mid := int.level]